jetson/requirements.txt
# === DEPENDENCIAS PRINCIPALES (requeridas para el proyecto) ===

# === LANGCHAIN CORE ===
langchain-core>=0.2.0,<0.3.0
langchain>=0.2.0,<0.3.0
langchain-community>=0.2.0,<0.3.0
langchain-huggingface>=0.0.1,<0.1.0

# === SPLITTERS & VECTORSTORES ===
langchain-text-splitters>=0.2.0,<0.3.0
langchain-chroma>=0.1.0,<0.2.0

# === Document Loaders ===
unstructured[markdown]>=0.15.0,<0.19.0

# === WebSocket + API + WebRTC ===
aiohttp>=3.9.0,<3.10.0
python-socketio[asyncio]>=5.11.0,<6.0.0
aiortc>=1.6.0,<2.0.0

# === MODELOS ===
transformers>=4.40.0,<4.45.0
torch>=2.0.0,<2.5.0
bitsandbytes>=0.41.0,<0.44.0
accelerate>=0.24.0,<0.34.0
sentence-transformers>=5.1.2,<6.0.0

# === PROCESAMIENTO DE IMAGEN/VIDEO ===
opencv-python>=4.9.0,<4.11.0
av>=12.0.0,<13.0.0

# === OTROS ===
python-dotenv>=1.0.0,<2.0.0
huggingface-hub>=0.20.0,<0.25.0
urllib3>=2.0.0,<2.3.0

#Jetson.GPIO
#RPi.GPIO